bdev: BdevRequest
Previously iovec is stack allocated. It turns out that SPDK bdev layer
does not copy the iovec but only keeps a pointer. When SPDK reads iovec
upon completion, it would cause undefined behavior.

Bdev_ReadPacket and Bdev_WritePacket now calculate blockCount from
pkt->pkt_len, and fills iovec total length to exactly
blockSize*blockCount. This appears to be a requirement in all SPDK
bdev drivers (in bdev_io_split), not just Malloc bdev.

fwdp can use file-backed bdev, in addition to Malloc.

+++ container/cs/test_test.go
@@ -97,7 +97,7 @@ func (f *Fixture) noError(e error) {
 }
 
 // EnableDisk enables on-disk caching on a Malloc Bdev.
-func (f *Fixture) EnableDisk(nSlots int) {
+func (f *Fixture) EnableDisk(nSlots int64) {
 	var e error
 	f.Bdev, e = bdev.NewMalloc(512, (1+nSlots)*16)
 	f.noError(e)
